Code Refactor for product update

+++ Priyosaj.Service/ProductService.cs
@@ -106,76 +106,101 @@ public class ProductService : IProductService
         await _unitOfWork.Complete();
     }
 
-    public async Task<ProductResponseDto> UpdateProductAsync(ProductUpdateReqDto product, string rootPath)
+    public async Task<ProductResponseDto> UpdateProductAsync(ProductUpdateReqDto productUpdateReq, string rootPath)
     {
         _currentUserService.ValidateIfEditor();
-        var spec = new ProductByIdSpecification(product.Id);
-
-        var productToUpdate = await _unitOfWork.Repository<Product>().GetEntityWithSpec(spec);
-
+        
+        var spec = new ProductByIdSpecification(productUpdateReq.Id);
+        var product = await _unitOfWork.Repository<Product>().GetEntityWithSpec(spec);
         if (product == null) throw new NotFoundException("Product not found");
-        productToUpdate.Title = product.Title;
-        productToUpdate.Description = product.Description;
-        productToUpdate.RegularPrice = product.RegularPrice;
-        productToUpdate.DiscountPrice = product.DiscountPrice;
 
-        if (product.DisplayImage != null)
+        var updatedProduct = _mapper.Map<Product>(productUpdateReq);
+
+        await HandleProductImagesUpdate(productUpdateReq, updatedProduct, rootPath);
+        HandleDisplayImageUpdate(productUpdateReq, updatedProduct);
+        await HandleProductCategoryUpdate(productUpdateReq, updatedProduct);
+        
+        _unitOfWork.Repository<Product>().Update(updatedProduct);
+
+        var result = await _unitOfWork.Complete();
+        if (result <= 0)
         {
-            productToUpdate.DisplayImageId = product.DisplayImage.Id;
-            var displayImageFile = productToUpdate.Images.FirstOrDefault(img => img.Id == product.DisplayImage.Id);
-            if (displayImageFile is null)
-            {
-                throw new NotFoundException("Display Image not found");
-            }
+            throw new BadRequestException("Product Update Failed!");
         }
 
-        var imagesToDelete = productToUpdate.Images.Where(img => !product.Images.Any(imgDto => imgDto.Id == img.Id)).ToList();
+        return _mapper.Map<ProductResponseDto>(updatedProduct);
+    }
 
+    private async Task HandleProductCategoryUpdate(ProductUpdateReqDto productUpdateReq, Product updatedProduct)
+    {
         var categories = new List<ProductCategory>();
-        foreach (var category in productToUpdate.ProductCategories)
+        foreach (var category in updatedProduct.ProductCategories)
         {
-            if (!product.ProductCategories.Contains(category.Id))
+            if (!productUpdateReq.ProductCategories.Contains(category.Id))
             {
                 categories.Add(category);
             }
         }
+
         foreach (var category in categories)
         {
-            productToUpdate.ProductCategories.Remove(category);
+            updatedProduct.ProductCategories.Remove(category);
         }
-        foreach (var id in product.ProductCategories)
+
+        foreach (var id in productUpdateReq.ProductCategories)
         {
             var category = await _unitOfWork.Repository<ProductCategory>().GetByIdAsync(id);
             if (category == null) throw new NotFoundException("Product Category not found");
-            if (productToUpdate.ProductCategories.Any(pc => pc.Id == category.Id))
+            if (updatedProduct.ProductCategories.Any(pc => pc.Id == category.Id))
             {
                 continue;
             }
-            productToUpdate.ProductCategories.Add(category);
+
+            updatedProduct.ProductCategories.Add(category);
             while (category.ParentId != null)
             {
                 category = await _unitOfWork.Repository<ProductCategory>().GetByIdAsync((Guid)category.ParentId);
-                productToUpdate.ProductCategories.Add(category);
+                updatedProduct.ProductCategories.Add(category);
             }
         }
-
-        await _fileUploadService.DeleteFiles(rootPath, imagesToDelete);
-
-        _unitOfWork.Repository<Product>().Update(productToUpdate);
-
-        var result = await _unitOfWork.Complete();
-
-        return _mapper.Map<ProductResponseDto>(productToUpdate);
     }
 
-    public async Task<ProductResponseDto> UploadImages(string productId, string webRootPath, IFormFileCollection images)
+    private async Task HandleProductImagesUpdate(ProductUpdateReqDto productUpdateReq, Product updatedProduct, string rootPath)
+    {
+        if (productUpdateReq.ImagesToDelete == null)
+        {
+            return;
+        }
+
+        if (updatedProduct.DisplayImageId != null && productUpdateReq.ImagesToDelete.Contains(updatedProduct.DisplayImageId.Value))
+        {
+            throw new BadRequestException("You cannot delete the display image of a product. Please select another image as your display image!");
+        }
+        
+        var imagesToDelete = updatedProduct.Images.Where(img => productUpdateReq.ImagesToDelete.Contains(img.Id)).ToList();
+        await _fileUploadService.DeleteFiles(rootPath, imagesToDelete);
+        updatedProduct.Images = updatedProduct.Images.Where(i => !productUpdateReq.ImagesToDelete.Contains(i.Id)).ToList();
+    }
+
+    private void HandleDisplayImageUpdate(ProductUpdateReqDto productUpdateReq, Product product)
+    {
+        if (productUpdateReq.DisplayImageToUpdate == null) return;
+        
+        var displayImageFile = product.Images.FirstOrDefault(img => img.Id == productUpdateReq.DisplayImageToUpdate.Id);
+        if (displayImageFile is null)
+        {
+            throw new NotFoundException("Display Image not found");
+        }
+        product.DisplayImageId = productUpdateReq.DisplayImageToUpdate.Id;
+    }
+
+    public async Task<ProductResponseDto> UploadImages(Guid productId, string webRootPath, IFormFileCollection images)
     {
         _currentUserService.ValidateIfEditor();
         var files = await _fileUploadService.UploadFiles("Product", webRootPath, images);
         try
         {
-            // Console.WriteLine(webRootPath);
-            var product = await _unitOfWork.Repository<Product>().GetByIdAsync(Guid.Parse(productId));
+            var product = await _unitOfWork.Repository<Product>().GetByIdAsync(productId);
             if (product == null) throw new NotFoundException("Product not found");
             foreach (var image in files)
             {
